MC-137: Add product associations export in product export code
