+++ Resources/config/batch_jobs.yml
@@ -11,12 +11,6 @@ connector:
                         reader:    pim_base_connector.reader.doctrine.product
                         processor: pim_magento_connector.processor.product_to_array
                         writer:    pim_magento_connector.writer.product
-                update_product_associations:
-                    title: pim_magento_connector.jobs.magento_product_associations_export.export.title
-                    services:
-                        reader:    pim_magento_connector.reader.product_association
-                        processor: pim_magento_connector.processor.associated_product_to_array
-                        writer:    pim_base_connector.writer.dummy
 
         magento_category_export:
             title: pim_magento_connector.jobs.magento_category_export.title
