+++ Resources/config/processors.yml
@@ -1,7 +1,6 @@
 parameters:
     pim_magento_connector.processor.product_to_array.class:            Pim\Bundle\MagentoConnectorBundle\Processor\ProductToArrayProcessor
     pim_magento_connector.processor.variant_group_to_array.class:      Pim\Bundle\MagentoConnectorBundle\Processor\VariantGroupToArrayProcessor
-    pim_magento_connector.processor.associated_product_to_array.class: Pim\Bundle\MagentoConnectorBundle\Processor\AssociatedProductToArrayProcessor
     pim_magento_connector.processor.attribute.class:                   Pim\Bundle\MagentoConnectorBundle\Processor\AttributeProcessor
     pim_magento_connector.processor.family.class:                      Pim\Bundle\MagentoConnectorBundle\Processor\FamilyProcessor
     pim_magento_connector.processor.add_attribute_to_sets.class:       Pim\Bundle\MagentoConnectorBundle\Processor\AddAttributeToSetsProcessor
@@ -12,6 +11,7 @@ services:
         class: %pim_magento_connector.processor.product_to_array.class%
         arguments:
             - '@pim_magento_connector.serializer'
+            - '@pim_catalog.manager.channel'
 
     pim_magento_connector.processor.attribute:
         class: %pim_magento_connector.processor.attribute.class%
@@ -32,12 +32,6 @@ services:
         class: %pim_magento_connector.processor.category.class%
         arguments:
             - '@pim_magento_connector.serializer'
-
-    pim_magento_connector.processor.associated_product_to_array:
-        class: %pim_magento_connector.processor.associated_product_to_array.class%
-        arguments:
-            - '@pim_magento_connector.serializer'
-            - '@pim_catalog.manager.channel'
 
     pim_magento_connector.processor.variant_group_to_array:
         class: %pim_magento_connector.processor.variant_group_to_array.class%
